app/requirements.txt
gunicorn==19.6.0
redis==2.10.5
Django==3.1
PyMySQL==0.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
mysqlclient
django-rest-swagger
django-filter==2.4.0
django-cors-headers
djangorestframework-camel-case
django_extensions
drf-extensions
uwsgi==2.0.17
django-debug-toolbar==3.1.1